requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn==0.24.0

# Data Validation and Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Data Processing
pandas==2.1.3

# Testing
pytest==7.4.3

# HubSpot API
hubspot-api-client==5.1.0

# Web Scraping & LLM
beautifulsoup4==4.12.2
lxml==5.3.0
crawl4ai==0.3.74
google-generativeai==0.3.0

# Auth, Database, and Payments
google-auth==2.28.1
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.25
stripe==8.8.0
